+++ common/lib/xmodule/xmodule/capa_base.py
@@ -94,7 +94,7 @@ class CapaFields(object):
         scope=Scope.settings,
         # it'd be nice to have a useful default but it screws up other things; so,
         # use display_name_with_default for those
-        default="Blank Advanced Problem"
+        default=_("Blank Advanced Problem")
     )
     attempts = Integer(
         help=_("Number of attempts taken by the student on this problem"),
@@ -425,9 +425,9 @@ class CapaMixin(CapaFields):
 
         # Apply customizations if present
         if 'custom_check' in self.text_customization:
-            check = _(self.text_customization.get('custom_check'))
+            check = _(self.text_customization.get('custom_check'))                # pylint: disable=translation-of-non-string
         if 'custom_final_check' in self.text_customization:
-            final_check = _(self.text_customization.get('custom_final_check'))
+            final_check = _(self.text_customization.get('custom_final_check'))    # pylint: disable=translation-of-non-string
         # TODO: need a way to get the customized words into the list of
         # words to be translated