+++ common/test/acceptance/tests/studio/test_studio_acid_xblock.py
@@ -10,6 +10,7 @@ from ...pages.xblock.acid import AcidView
 from ...fixtures.course import CourseFixture, XBlockFixtureDesc
 
 
+@skip('Jenkins builds are getting stuck on acid_block tests')
 class XBlockAcidBase(WebAppTest):
     """
     Base class for tests that verify that XBlock integration is working correctly
@@ -116,6 +117,7 @@ class XBlockAcidNoChildTest(XBlockAcidBase):
         self.user = course_fix.user
 
 
+@skip('Jenkins builds are getting stuck on acid_block tests')
 class XBlockAcidParentBase(XBlockAcidBase):
     """
     Base class for tests that verify that parent XBlock integration is working correctly
@@ -169,6 +171,7 @@ class XBlockAcidEmptyParentTest(XBlockAcidParentBase):
         self.user = course_fix.user
 
 
+@skip('Jenkins builds are getting stuck on acid_block tests')
 class XBlockAcidChildTest(XBlockAcidParentBase):
     """
     Tests of an AcidBlock with children